+++ src/DCC.cpp
@@ -788,7 +788,7 @@ void DCC::setCBRGCallback (std::function<void()> cb)
     m_cbr_g_callback = std::move(cb);
 }
 
-void DCC::updateAoI (int priority, int64_t time)
+void DCC::updateAoI (int priority, double time)
 {
     switch(priority)
     {
